Created a redux slice named ItemPassedToModalSlice with two reducers that stores the item chosen from the dashboardMenu component into an array, and clears the array. Updated redux store to include the new slice.
Updated the dashboardMenu component to use the new itemPassedToModal slice to pass a selected menu item to the menuItemModal component after an onClick function.

Added a button to the mennuItemModal component which passes the array from the itemPassedToModal slice to the activOrder slice in the order component, then runs the second reducer of the ItemPassedToModal reducers to return the itemPassedToModal state to an empty array.

Made changes in the order component to the mapped array to correctly reflect the array structure.