Amended the currentOrder object in the ActiveOrderSlice to accept a sedond value within the data array which contains the quantity of an item selected in the MenuItemModal component.
Amended the addItemToActiveOrder function in the MenuItemModal component to pass in the item to add to the active order, as well as the item quantity selected.

In the MenuItemModal component, renamed SVG add and remove elements' onClick functions.

Imported the sumOrderPrice reducer from the OrderPriceSlice to calculate the price of an item times the quantity chosen. Changed the OrderPrice reducer to SumOrderPrice and refactored the code to calculate the price times quantity passed from the MenuItemModal component. Also created a deductRemovedItemPrice reducer in the same slice which reduces the payload data from the orderPrice state.

Imported the deductRemovedItemPrice to the order component, and added the reducer in the handleClick function in the order component when removing an item from the activeOrder which deducts the price times the quantity from the orders total price.

Html changes in the order component to reflect the structure changes in the ActiveOrderSlice which changed from accepting a single array with an object, to the same array with a quantity value in the next index.

Imported the sumOrderPrice reducer to the price component, and initialized the orderPrice state to a const variable. Temporarily updated the complete order onClick function to log the orderPrice and made a minor html change to the total price displayed to show two decimal places.