Created a displayDeliveryOptionScreen slice in redux with a single reducer that manages a boolean value, updated the redux store action / reducer.
Created a DeliveryOptionScreen component screen component which is the first component to render on the dashboard component. Used the logical AND operator to render the DeliveryOptionScreen component in the Dashboard component if the boolean value is true. Used the logical AND operator to render the dashboard child components if the value is false.

Created and styled two buttons in the DeliveryOptionScreen component which both run the displayDeliveryOptionScreen reducer through onClick functions setting the state to false which unrenders the DeliveryOptionScreen component and renders the remaining child components of the dashboard component.

Created a css class selector for buttons in the main index.css file in order to change the cursor to pointer on element hover.