Added in missing alt text from image elements.
Created a MenuNavCategorySlice in redux to hold the state of the MenuNav component and imported to redux store.

Removed each individual DashboardCard component, imported the MenuNavCategorySlice state and mapped a single DashboardCard component, passing the state data as props into the DashboardCard component.

Refactored the DashboardCard component's html elements and restyled the rendered button.