Created a new redux slice to record whether the order is a delivery or not. Imported the new slice to the redux store.
Imported the new state into the Price component and two conditionals to display whether a delivery fee is added to the order based on the delivery state. If not, no additional fee is added. Text added to reflect whether the order is delivery or not and rendered based on the conditional.

Minor changes made to the order component and an unused import removed, as well as minor changes to the component styling.

In the DeliveryOptionScreen component, added in a title and text as well as changed the styling of the component. Imported the new delivery slice from redux and included it in the onClick function for delivery selection.

Minor changes to the styling of the DashboardMenu component styling.