created a scroll class and a conditional in the order component to only render the scroll bar on overflow.
Created and styled a basic ConfirmationModal component for when an order is confirmed and user is ready to check out. Updated the new component import and exports. Included a handleClose onClick function which renders the modal inactive.

In the Price component, imported useDispatch, and amended the onClick function for the order button to render the ConfirmationModal component.

Created a new slice in redux which records the active or inactive state of the new order modal. Updated the redux store import.

In the dashboard component, imported the new confirmation modal component as well as the active or inactive state to render the component, then used short circuiting with the logical AND operator to render the confirmation modal component if the new order modal state is true, and unrender if false.