Created a DeliveryForm component and NPM installed React Google Autocomplete. Created a basic form which includes an Autocomplete component and a button. When an address is searched in the Autocomplete component, useState is used to store the address locally. The button has an onClick function which has a conditional run if the address length is greater than 0; then the adress state is passed into redux through a new slice in redux called AddressSlice.
The new addressSlice created takes a string as the state, and the reducer replaces the string when the reducer is ran. Redux store also updated to reflect new slice.

Amended the DeliveryOption component and moved the button elements for delivery or pickup options to a newly created component called DeliveyOptionScreen. Created a useState in the component named displayDeliveryComponent which takes a boolean value. Wrote a conditional in the component which renders either the DeliveryForm or the DeliveryOption component based on the useState value. In the same component, temporarily prop drilled setDisplayDeliveryComponent to DeliveryForm for testing which will be removed later. Prop drilled the delivery / pickup onClick functions to DeliveryOption to match up with the buttons which had moved components.

All newly created dashboard components imported to the components index.js file and exported appropriately. Basic styling added to newly created and amended components.

New Dotenv file created and the Google API key added to the .env file. .env file types added to gitignore.